Remove adapters workaround
Previous versions of lawnchair accepted only one adapter as parameter
on the constructor. For that reason we were using a workaround to allow
us use the best adapter we have available (the code was actually stolen
from lawnchair's master branch). Since we had to use the trunk version
anyway (due to problems with the indexed-db adapter on version 0.6.1),
we don't need this workaround anymore.

Signed-off-by: Mailson Menezes <mailson@gmail.com>